+++ lib/src/utils/utils.dart
@@ -85,28 +85,33 @@ FieldElement? findTransferIdField(ClassElement element) {
 extension DartTypeExtension on DartType {
   /// Checks if [type] is a [TransferObject] annotated class.
   bool get isTransferObject {
-    if (element2 == null) {
+    if (element == null) {
       return false;
     }
 
     return TypeChecker.fromRuntime(TransferObject)
-            .firstAnnotationOfExact(element2!, throwOnUnresolved: false) !=
+            .firstAnnotationOfExact(element!, throwOnUnresolved: false) !=
         null;
   }
 
   /// Checks if [type] is a [Hub] annotated class.
   bool get isHub {
-    if (element2 == null) {
+    if (element == null) {
       return false;
     }
 
     return TypeChecker.fromRuntime(Hub)
-            .firstAnnotationOfExact(element2!, throwOnUnresolved: false) !=
+            .firstAnnotationOfExact(element!, throwOnUnresolved: false) !=
         null;
   }
 
-  bool get isResource =>
-      TypeChecker.fromRuntime(Resource).isAssignableFromType(this);
+  static const _resourceBaseTypes = [
+    Resource,
+    CollectionResource,
+  ];
+
+  bool get isResource => _resourceBaseTypes
+      .any((t) => TypeChecker.fromRuntime(t).isAssignableFromType(this));
 
   bool get isDartCoreDateTime =>
       TypeChecker.fromRuntime(DateTime).isExactlyType(this);
@@ -114,7 +119,7 @@ extension DartTypeExtension on DartType {
   bool get isDartCoreDuration =>
       TypeChecker.fromRuntime(Duration).isExactlyType(this);
 
-  bool get isEnum => (element2 is ClassElement) && (element2 is EnumElement);
+  bool get isEnum => (element is ClassElement) && (element is EnumElement);
 
   bool get isUint8List =>
       TypeChecker.fromRuntime(Uint8List).isExactlyType(this);
@@ -167,4 +172,14 @@ DartType? readTypeField(DartObject? element, String fieldName) {
 String getLayoutName(ClassElement element) {
   final annotation = getAnnotation(element, DaoType);
   return readField<String>(annotation, 'name') ?? element.name.toLowerCase();
+}
+
+extension StringExtension on String {
+  String get firstUpper {
+    if (isEmpty) {
+      return this;
+    }
+
+    return substring(0, 1).toUpperCase() + substring(1);
+  }
 }